refactor: move `MessageSendFailureHandler` to `feature/message`
